dighosp-frontend/pyproject.toml
[tool.poetry]
name = "dighosp-frontend"
version = "0.1.0a0"
description = ""
authors = ["Yin-Chi Chan <ycc39@cam.ac.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
dash = "^2.17.1"
dash-bootstrap-components = "^1.6.0"
dash-compose = "^2023.3.12"
dash-ag-grid = "^31.2.0"
plotly = "^5.22.0"
numpy = "^2.0.0"
pandas = "^2.2.2"
humanize = "^4.10.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
toml = "^0.10.2"
orjson = "^3.10.6"


[tool.poetry.group.dev.dependencies]
pylint = "^3.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 100
color_output = true
known_first_party = ["dighosp-frontend"]

[tool.autopep8]
max_line_length = 100
